Removed use of tempfile persist.
For users who have their project on a different storage device to their
temp directory it causes a panic.

The code creates a temporary file, then persists it to the data
directory only to remove it within the test. This seems superfluous.

The current setup just uses the tempfile where it stands and relies on
the destructors in tempfile to remove files when they fall out of scope.

I have tested the Solution tests on OSX, Windows and Linux. All passed.